FIX: cast input to float

+++ code/train.py
@@ -132,8 +132,9 @@ def main(unused_argv):
                             nresblocks=params.nresblocks, dim=params.depth)
         discriminator = partial(network.discriminator, scope=constants.dis_scope, dim=params.depth)
 
-        hr_images_real = {factor: params.hr_scale * downscale2d(hr_image, factor=factor) for factor in constants.factors}
-        hr_images_fake = generator(params.lr_scale * lr_image)
+        hr_images_real = {factor: params.hr_scale * tf.cast(downscale2d(hr_image, factor=factor), tf.float32) 
+                          for factor in constants.factors}
+        hr_images_fake = generator(params.lr_scale * tf.cast(lr_image, tf.float32))
 
         # model outputs
         gen = {factor: (1.0 / params.hr_scale) * hr_images_fake[factor] for factor in constants.factors}
